Data for new modeler (4.1) : separate linear problem data implementation from its abstract parent
